requirements.txt
flask==2.3.3
google-generativeai==0.3.1
beautifulsoup4==4.12.2
requests==2.31.0
yfinance==0.2.31
pandas==2.1.1
python-dotenv==1.0.0
fpdf2==2.7.6
langchain==0.0.312
langchain-google-genai==0.0.5
playwright==1.40.0
jinja2==3.1.2 